Refactored utils and individual commands
